add: fix a minor bug in the no 2 game

+++ Soal_2_AryoWisanggeni_13523100_Cakru17/classes.py
@@ -232,7 +232,7 @@ class Game:
         if (newPassive == "speed malfunction"):
             newRobot = Robot(newName, newMaxHP, newAttackPower, newSpeed, SpeedMalfunction())
         if (newPassive == "last upgrade"):
-            attack_multi = int(input("Enter attack multiplication value: "))
+            attack_multi = float(input("Enter attack multiplication value: "))
             newRobot = Robot(newName, newMaxHP, newAttackPower, newSpeed, LastUpgrade(attack_multi))
 
         self.robotList.append(newRobot)
